dom/tests/mochitest/browser-frame/test_browserFrame10.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=753595
-->
 <head>
  <title>Test for Bug 753595</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserFrameHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=753595">Mozilla Bug 753595</a>

<!--
   Test the getScreenshot property for mozbrowser
-->

<script type="application/javascript;version=1.7">

SimpleTest.waitForExplicitFinish();

var iframeScript = function() {
  content.document.addEventListener("mozvisibilitychange", function() {
    sendAsyncMessage('test:visibilitychange', {
      hidden: content.document.mozHidden
    });
  }, false);
}

function runTest() {

  browserFrameHelpers.setEnabledPref(true);
  browserFrameHelpers.addToWhitelist();

  var iframe1 = document.createElement('iframe');
  iframe1.mozbrowser = true;
  iframe1.src = 'data:text/html,1';

  document.body.appendChild(iframe1);

  function recvVisibilityChanged(msg) {
    ok(true, 'iframe recieved visibility changed');
    ok(msg.json.hidden === true, 'mozHidden attribute correctly set');
    SimpleTest.finish();
  }

  function iframeLoaded() {
    var mm = SpecialPowers.getBrowserFrameMessageManager(iframe1);
    mm.addMessageListener('test:visibilitychange', recvVisibilityChanged);
    mm.loadFrameScript('data:,(' + iframeScript.toString() + ')();', false);
    info('setting invactive');
    iframe1.setActive(false).onsuccess = function(e) {}
  }

  iframe1.addEventListener('mozbrowserloadend', iframeLoaded);
}

addEventListener('load', function() { SimpleTest.executeSoon(runTest); });

</script>

</body>
</html>
